Initial refactor to make the code more readable.
