+++ tests/test_init.py
@@ -61,7 +61,7 @@ class TestFlumeAuth(unittest.TestCase):
             mock: Requests mock.
 
         """
-        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
+        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.constants.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
         auth = pyflume.FlumeAuth(
             CONST_USERNAME, CONST_PASSWORD, CONST_CLIENT_ID, CONST_CLIENT_SECRET, CONST_FLUME_TOKEN, http_session=Session(),
         )
@@ -81,10 +81,10 @@ class TestFlumeDeviceList(unittest.TestCase):
             mock: Requests mock.
 
         """
-        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
+        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.constants.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
         mock.register_uri(
             'get',
-            pyflume.API_DEVICES_URL.format(user_id=CONST_USER_ID),
+            pyflume.constants.API_DEVICES_URL.format(user_id=CONST_USER_ID),
             text=load_fixture('devices.json'),
         )
         flume_auth = pyflume.FlumeAuth(
@@ -110,10 +110,10 @@ class TestFlumeNotificationList(unittest.TestCase):
             mock: Requests mock.
 
         """
-        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
+        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.constants.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
         mock.register_uri(
             'get',
-            pyflume.API_NOTIFICATIONS_URL.format(user_id=CONST_USER_ID),
+            pyflume.constants.API_NOTIFICATIONS_URL.format(user_id=CONST_USER_ID),
             text=load_fixture('notification.json'),
         )
         flume_auth = pyflume.FlumeAuth(
@@ -139,10 +139,10 @@ class TestFlumeData(unittest.TestCase):
             mock: Requests mock.
 
         """
-        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
+        mock.register_uri(CONST_HTTP_METHOD_POST, pyflume.constants.URL_OAUTH_TOKEN, text=load_fixture(CONST_TOKEN_FILE))
         mock.register_uri(
             CONST_HTTP_METHOD_POST,
-            pyflume.API_QUERY_URL.format(user_id=CONST_USER_ID, device_id='device_id'),
+            pyflume.constants.API_QUERY_URL.format(user_id=CONST_USER_ID, device_id='device_id'),
             text=load_fixture('query.json'),
         )
